requirements.txt
pandas
numpy 
python-dotenv
mysql-connector-python
pymysql
scikit-learn

-e .
